Patch for addNeighbour cause model modified for social & aboutMeTxt

+++ app/src/main/java/com/openclassrooms/entrevoisins/model/Neighbour.java
@@ -9,7 +9,7 @@ import java.util.Objects;
 public class Neighbour implements Serializable {
 
     /** Identifier */
-    private Integer id;
+    private Long id;
 
     /** Full name */
     private String name;
@@ -50,7 +50,7 @@ public class Neighbour implements Serializable {
      * @param name
      * @param avatarUrl
      */
-    public Neighbour(Integer id, String name, String avatarUrl, String address,
+    public Neighbour(Long id, String name, String avatarUrl, String address,
                      String phoneNumber, String social, String aboutMe, String aboutMeText) {
         this.id = id;
         this.name = name;
@@ -62,11 +62,11 @@ public class Neighbour implements Serializable {
         this.aboutMeText = aboutMeText;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
